+++ src/main/java/cart/controller/PageController.java
@@ -1,0 +1,36 @@
+package cart.controller;
+
+import cart.dto.ProductResponse;
+import cart.service.ProductService;
+import java.util.List;
+import java.util.stream.Collectors;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class PageController {
+    private final ProductService productService;
+
+    public PageController(ProductService productService) {
+        this.productService = productService;
+    }
+
+    @GetMapping("/")
+    public String index(Model model) {
+        List<ProductResponse> products = productService.findProducts().stream()
+                .map(ProductResponse::from)
+                .collect(Collectors.toList());
+        model.addAttribute("products", products);
+        return "index.html";
+    }
+
+    @GetMapping("/admin")
+    public String admin(Model model) {
+        List<ProductResponse> products = productService.findProducts().stream()
+                .map(ProductResponse::from)
+                .collect(Collectors.toList());
+        model.addAttribute("products", products);
+        return "admin.html";
+    }
+}
