+++ src/main/java/cart/controller/ProductController.java
@@ -1,48 +1,25 @@
 package cart.controller;
 
 import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
 import cart.entity.Product;
 import cart.service.ProductService;
 import java.net.URI;
-import java.util.List;
-import java.util.stream.Collectors;
 import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
+@RestController
 public class ProductController {
 
     private final ProductService productService;
 
     public ProductController(ProductService productService) {
         this.productService = productService;
-    }
-
-    @GetMapping("/")
-    public String index(Model model) {
-        List<ProductResponse> products = productService.findProducts().stream()
-                .map(ProductResponse::from)
-                .collect(Collectors.toList());
-        model.addAttribute("products", products);
-        return "index";
-    }
-
-    @GetMapping("/admin")
-    public String admin(Model model) {
-        List<ProductResponse> products = productService.findProducts().stream()
-                .map(ProductResponse::from)
-                .collect(Collectors.toList());
-        model.addAttribute("products", products);
-        return "admin";
     }
 
     @PostMapping("/products")
@@ -52,7 +29,7 @@ public class ProductController {
     }
 
     @PatchMapping("/products/{id}")
-    public ResponseEntity<Void> updateProduct(@PathVariable long id,
+    public ResponseEntity<Void> updateProduct(@PathVariable Long id,
                                               @Valid @RequestBody ProductRequest productRequest) {
         Product product = new Product(id, productRequest.getName(), productRequest.getImgUrl(),
                 productRequest.getPrice());
@@ -61,7 +38,7 @@ public class ProductController {
     }
 
     @DeleteMapping("/products/{id}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable long id) {
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
         productService.deleteProductBy(id);
         return ResponseEntity.noContent().build();
     }
